Fix key vault secret name mismatch
The key vault secret name for the database user and admin user in the main.parameters.json did not match the secret name in the [flexibleserver|sqlserver].bicep files.

The latter files are the one that save the passwords into the secrets, and because the secrets never get saved by the name that the main.parameters.json specifies, it kept generatign new passwords on each deployment. This resulted in the connection string getting invalid starting with the second deployment.